requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
openai==1.50.0
anthropic==0.34.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
redis==5.1.0
boto3==1.35.0
python-multipart==0.0.9
python-dotenv==1.0.1
websockets==12.0
prometheus-client==0.21.0
structlog==24.4.0
tenacity==9.0.0
docker==7.1.0
Pillow==10.4.0
